requirements.txt
pandas>=2.0
numpy>=1.24
scikit-learn>=1.3
pybaseball>=2.2.6
matplotlib>=3.7
seaborn>=0.13
pyyaml>=6.0
tqdm>=4.66
streamlit>=1.33

pytest>=7.0
